Refactored endpoints into their separate files
